+++ packages/flow-table/CHANGELOG.md
@@ -3,6 +3,12 @@
 
 # Change Log
 
+## [1.3.2] - 2023-09-13
+
+### Bug Fixes
+
+- `f-table-schema` : cell with template not taking full width.
+
 ## [1.3.1] - 2023-09-13
 
 ### Improvements